+++ test/resources/delete-local-browser.js
@@ -5,7 +5,7 @@ const { expect } = require("chai");
 var localDB = null
 
 
-describe("Delete resource tests", function(){
+describe("Delete local browser tests", function(){
 
     this.timeout(30000)
 
@@ -57,15 +57,17 @@ describe("Delete resource tests", function(){
         expect(del && found == false).to.be.true
     })
 
-    it("should return true if a resource is deleted and the success notification appeared", async () => {
-        var del = null
-        var notification = null;
-
-        del = await utils.resources.deleteResource("test")
-        
-        await utils.sleep(1000)
-
-        notification = await utils.checkSuccessNotification()
-        expect(del && notification).to.be.true
-    })
+    if(global.env == "w"){
+        it("should return true if a resource is deleted and the success notification appeared", async () => {
+            var del = null
+            var notification = null;
+    
+            del = await utils.resources.deleteResource("test")
+            
+            await utils.sleep(1000)
+    
+            notification = await utils.checkSuccessNotification()
+            expect(del && notification).to.be.true
+        })
+    }
 })
